move edit merchant fuctionality to its own spec file and add to not have testing
